docs(test): scope Issue #7 as partial local validation only
- Phase 1 checklist: leave staging/production item open; clarify local test DB
  and integration tests are not deploy sign-off; track #9 for product BPMN gaps.
- BPMN_STAGING_VALIDATION: reframe as partial evidence; rename optional tests to
  diagnostic module with ISSUE7_PRODUCT_BPMN_DIAGNOSTIC; tighten completion
  asserts only for bi_weekly_report and property_due_diligence; observability-only
  for bidder_onboarding and dap_report until #9.

Refs #7, #9

Made-with: Cursor
